Move Effects into //lib/effect
This is the start of normalizing our repo structure to follow PIO
standards.
