+++ lib/effect/ColorCycleEffect.hpp
