+++ lib/effect/ContrastBumpsEffect.hpp
