+++ lib/effect/ControlEffect.cpp
