+++ lib/effect/DarkEffect.cpp
