+++ lib/effect/DisplayColorPaletteEffect.cpp
