+++ lib/effect/DisplayColorPaletteEffect.hpp
