+++ lib/effect/Effect.cpp
