+++ lib/effect/FireflyEffect.cpp
