+++ lib/effect/LightningEffect.hpp
