+++ lib/effect/PoliceEffect.hpp
