+++ lib/effect/RainbowBumpsEffect.hpp
