+++ lib/effect/RainbowEffect.cpp
