+++ lib/effect/RainbowEffect.hpp
