+++ lib/effect/RorschachEffect.hpp
