+++ lib/effect/SimpleBlinkEffect.cpp
