+++ lib/effect/SimpleBlinkEffect.hpp
