+++ lib/effect/SparkEffect.hpp
