+++ lib/effect/StopLightEffect.cpp
