+++ lib/effect/StopLightEffect.hpp
