+++ lib/effect/SwingingLights.hpp
