+++ software/arduino/FastLedManager.hpp
@@ -4,7 +4,8 @@
 #include <Types.hpp>
 #include <vector>
 
-#include "../generic/Effect.hpp"
+#include "../../software/generic/DeviceDescription.hpp"
+#include "../../lib/effect/Effect.hpp"
 #include "../generic/LedManager.hpp"
 
 class FastLedManager : public LedManager {
